Views & more models logic (not finished)

+++ core/src/main/java/com/krolewskie_potyczki/model/SquareUnit.java
@@ -1,7 +1,7 @@
 package com.krolewskie_potyczki.model;
 
 public class SquareUnit extends Unit {
-    SquareUnit(boolean isPlayersEntity, double x, double y) {
+    SquareUnit(boolean isPlayersEntity, float x, float y) {
         super(isPlayersEntity, x, y);
     }
